quelques modifs pour que ca compile

+++ src/cli/compression/quantization.h
@@ -2,9 +2,10 @@
 #define DEF_QUANTIZATION
 
 #include <stdlib.h>
+#include <stdint.h>
 #include <math.h>
 
 void *quantizate(float *x, size_t *n, const float factor, int *nbbits);
-float *dequantizate(float *x, size_t *n, const float factor, int nbbits);
+float *dequantizate(uint8_t *x, size_t *n, const float factor, int nbbits);
 
 #endif
